+++ opengever/base/exceptions.py
@@ -1,0 +1,3 @@
+class TransportationError(Exception):
+    """A object could not be transported from or to another client.
+    """
