+++ opengever/base/transport.py
@@ -1,6 +1,6 @@
 from five import grok
-from opengever.ogds.base.exceptions import TransportationError
-from opengever.ogds.base.interfaces import IDataCollector
+from opengever.base.exceptions import TransportationError
+from opengever.base.interfaces import IDataCollector
 from opengever.ogds.base.utils import decode_for_json
 from opengever.ogds.base.utils import encode_after_json
 from opengever.ogds.base.utils import remote_json_request
