+++ opengever/inbox/browser/refuse.py
@@ -1,10 +1,10 @@
 from five import grok
+from opengever.base.transport import ORIGINAL_INTID_ANNOTATION_KEY
+from opengever.base.transport import REQUEST_KEY
+from opengever.base.transport import Transporter
 from opengever.inbox import _
 from opengever.inbox.browser.schema import ISimpleResponseForm
 from opengever.inbox.forwarding import IForwarding
-from opengever.ogds.base.transport import ORIGINAL_INTID_ANNOTATION_KEY
-from opengever.ogds.base.transport import REQUEST_KEY
-from opengever.ogds.base.transport import Transporter
 from opengever.ogds.base.utils import get_current_org_unit
 from opengever.ogds.base.utils import ogds_service
 from opengever.ogds.base.utils import remote_json_request
@@ -20,7 +20,6 @@ from z3c.form import button
 from z3c.form.field import Fields
 from z3c.form.form import Form
 from zope.annotation.interfaces import IAnnotations
-from zope.component import getUtility
 from zope.event import notify
 from zope.lifecycleevent import ObjectModifiedEvent
 import json
