+++ opengever/ogds/base/exceptions.py
@@ -1,8 +1,3 @@
 class ClientNotFound(Exception):
     """The client could not be found.
     """
-
-
-class TransportationError(Exception):
-    """A object could not be transported from or to another client.
-    """
