+++ opengever/task/browser/accept/utils.py
@@ -1,9 +1,9 @@
 from five import grok
+from opengever.base.transport import Transporter
 from opengever.base.utils import ok_response
 from opengever.globalindex.model.task import Task
 from opengever.inbox.utils import get_current_inbox
 from opengever.inbox.yearfolder import get_current_yearfolder
-from opengever.ogds.base.transport import Transporter
 from opengever.ogds.base.utils import get_current_admin_unit
 from opengever.ogds.base.utils import get_current_org_unit
 from opengever.ogds.base.utils import remote_request
