+++ opengever/task/transporter.py
@@ -1,9 +1,9 @@
 from DateTime import DateTime
 from datetime import datetime
 from five import grok
+from opengever.base.transport import ORIGINAL_INTID_ANNOTATION_KEY
+from opengever.base.transport import Transporter
 from opengever.base.utils import ok_response
-from opengever.ogds.base.transport import ORIGINAL_INTID_ANNOTATION_KEY
-from opengever.ogds.base.transport import Transporter
 from opengever.ogds.base.utils import remote_json_request
 from opengever.ogds.base.utils import remote_request
 from opengever.task.adapters import IResponse as IPersistentResponse
